feat(statistic): record the interval between consecutive frames
UpdateTimeAspect measures how long each GameSystem takes to run. That is
CPU work: a frame that starts late still reports the same per-system
durations, so a loop dropping from 20 FPS to 10 FPS leaves every existing
metric unchanged. Nothing recorded the interval between frames.

Add FrameIntervalAspect, which advises FrameDataSystem.earlyUpdate - the
first statement of WordOnlineLoop.update(), so it runs exactly once per
frame - and records the gap since the previous frame under the name
"Frame". The loop's own update() is package-private and cannot be advised
by Spring AOP, and routing this through an aspect keeps the game package
free of any dependency on statistic.

The previous timestamp lives on GameResultBuilder rather than in a map
owned by the aspect, so it is discarded together with the builder when
saveGameResult removes it instead of leaking an entry per game played.

Supporting changes:

- Widen UpdateTimeStatistic min/max and addInterval to long. The old int
  cast capped intervals at ~2.147s, which never mattered for microsecond
  per-system durations but would corrupt exactly the stalled frames worth
  measuring.
- Key the statistic map by String instead of Class<? extends GameSystem>,
  since a frame-interval entry is not a GameSystem. StatisticService keeps
  its Class parameter and resolves the simple name internally, so rows
  already written for each system keep an identical name value and no
  schema change is needed.
- Add the statistic_update_time definition to the test schema. It existed
  nowhere in the repository except the INSERT statement, so the
  persistence path could not be tested at all.

+++ src/main/java/com/wordonline/server/statistic/domain/UpdateTimeStatistic.java
@@ -5,12 +5,12 @@ import lombok.Getter;
 @Getter
 public class UpdateTimeStatistic {
 
-    private int minInterval = Integer.MAX_VALUE;
-    private int maxInterval = Integer.MIN_VALUE;
+    private long minInterval = Long.MAX_VALUE;
+    private long maxInterval = Long.MIN_VALUE;
     private float meanInterval;
     private int frameNum = 0;
 
-    public void addInterval(int interval) {
+    public void addInterval(long interval) {
         minInterval = Math.min(minInterval, interval);
         maxInterval = Math.max(maxInterval, interval);
